restricted pages added

+++ base/views.py
@@ -2,14 +2,18 @@ from django.shortcuts import render, redirect
 from django.contrib import messages
 from .models import Room, Topic
 from django.db.models import Q
+from django.http import HttpResponse
 from .forms import RoomForm
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
-
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username=request.POST.get('username')
@@ -50,6 +54,7 @@ def room(request ,pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -67,6 +72,9 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
+
     if request.method=='POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -78,6 +86,10 @@ def updateRoom(request, pk):
 
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
+
     if request.method=="POST":
         room.delete()
         return redirect('home')
